@@ -1,18 +1,17 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 24
+    compileSdkVersion 30
 
 
     defaultConfig {
         applicationId "com.azazel.cafecrawler"
-        minSdkVersion 15
-        targetSdkVersion 26
+        minSdkVersion 24
+        targetSdkVersion 30
         versionCode 38
         versionName "38"
 
-
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
     }
     buildTypes {
@@ -35,10 +34,10 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:24.2.1'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
     implementation project(':framework')
 
-    implementation 'com.android.support:design:24.2.1'
+    implementation 'com.google.android.material:material:1.0.0'
     implementation 'com.etsy.android.grid:library:1.0.5'
     implementation 'com.baoyz.pullrefreshlayout:library:1.0.1'
     implementation 'org.jsoup:jsoup:1.8.1'
