@@ -1,8 +1,7 @@
 package com.azazel.cafecrawler;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 
 import com.azazel.cafecrawler.fragment.ICrawlFragment;
 import com.azazel.framework.view.SmartFragmentStatePagerAdapter;
