@@ -3,13 +3,14 @@ package com.azazel.cafecrawler.fragment;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.graphics.Bitmap;
-import android.support.v7.app.AlertDialog;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.appcompat.app.AlertDialog;
 
 import com.azazel.cafecrawler.AlarmManager;
 import com.azazel.cafecrawler.CrawlConstants;
